requirements.txt
python-telegram-bot==21.1
urllib3==1.26.18
pytz
APScheduler
